GH-49434: [C++][CI] Add golden integration files to IPC file fuzz corpus

+++ cpp/build-support/fuzzing/generate_corpuses.sh
@@ -29,7 +29,7 @@ set -ex
 CORPUS_DIR=/tmp/corpus
 PANDAS_DIR=/tmp/pandas
 
-ARROW_ROOT=$(cd $(dirname $BASH_SOURCE)/../../..; pwd)
+ARROW_ROOT=$(cd $(dirname "$BASH_SOURCE")/../../..; pwd)
 ARROW_CPP=$ARROW_ROOT/cpp
 OUT=$1
 
@@ -39,10 +39,11 @@ OUT=$1
 
 # Arrow IPC
 
-IPC_INTEGRATION_FILES=$(find ${ARROW_ROOT}/testing/data/arrow-ipc-stream/integration -name "*.stream")
-
 rm -rf ${CORPUS_DIR}
 ${OUT}/arrow-ipc-generate-fuzz-corpus -stream ${CORPUS_DIR}
+# Add "golden" IPC integration files
+IPC_INTEGRATION_FILES=$(find ${ARROW_ROOT}/testing/data/arrow-ipc-stream/integration -name "*.stream")
+[ -z "${IPC_INTEGRATION_FILES}" ] && exit 1
 # Several IPC integration files can have the same name, make sure
 # they all appear in the corpus by numbering the duplicates.
 cp --backup=numbered ${IPC_INTEGRATION_FILES} ${CORPUS_DIR}
@@ -50,6 +51,9 @@ ${ARROW_CPP}/build-support/fuzzing/pack_corpus.py ${CORPUS_DIR} ${OUT}/arrow-ipc
 
 rm -rf ${CORPUS_DIR}
 ${OUT}/arrow-ipc-generate-fuzz-corpus -file ${CORPUS_DIR}
+IPC_INTEGRATION_FILES=$(find ${ARROW_ROOT}/testing/data/arrow-ipc-stream/integration -name "*.arrow_file")
+[ -z "${IPC_INTEGRATION_FILES}" ] && exit 1
+cp --backup=numbered ${IPC_INTEGRATION_FILES} ${CORPUS_DIR}
 ${ARROW_CPP}/build-support/fuzzing/pack_corpus.py ${CORPUS_DIR} ${OUT}/arrow-ipc-file-fuzz_seed_corpus.zip
 
 rm -rf ${CORPUS_DIR}
